+++ History.md
@@ -1611,3 +1611,17 @@ commit by afterloe (lm6289511@gmail.com)
 > services/userService.js  
 > tools/chain.js  
 
+
+1.0.3 / 2017-2-12 16:29:1
+==================
+commit by afterloe (lm6289511@gmail.com)
+
+  * master: ws职责链调优
+
+
+### modify files
+> servers/lib/ws-handlerChain/echoChain.js  
+> servers/lib/ws-handlerChain/index.js  
+> servers/lib/ws-handlerChain/managerChain.js  
+> servers/websocket.js  
+